+++ mono-traversable/src/Data/Sequences.hs
@@ -23,15 +23,12 @@ import Control.Category
 import Control.Arrow ((***), first, second)
 import Control.Monad (liftM)
 import qualified Data.Sequence as Seq
-import qualified Data.DList as DList
 import qualified Data.Vector as V
 import qualified Data.Vector.Unboxed as U
 import qualified Data.Vector.Storable as VS
 import Data.String (IsString)
 import qualified Data.List.NonEmpty as NE
 import qualified Data.ByteString.Unsafe as SU
-import Data.GrowingAppend
-import Data.Vector.Instances ()
 import qualified Data.Vector.Generic as VG
 import qualified Data.Vector.Algorithms.Merge as VAM
 import Data.Ord (comparing)
@@ -926,30 +923,6 @@ instance IsSequence (Seq.Seq a) where
     {-# INLINE index #-}
     {-# INLINE indexEx #-}
     {-# INLINE unsafeIndex #-}
-
-instance SemiSequence (DList.DList a) where
-    type Index (DList.DList a) = Int
-    cons = DList.cons
-    snoc = DList.snoc
-
-    reverse = defaultReverse
-    sortBy = defaultSortBy
-    intersperse = defaultIntersperse
-    find = defaultFind
-    {-# INLINE intersperse #-}
-    {-# INLINE reverse #-}
-    {-# INLINE find #-}
-    {-# INLINE sortBy #-}
-    {-# INLINE cons #-}
-    {-# INLINE snoc #-}
-
-instance IsSequence (DList.DList a) where
-    fromList = DList.fromList
-    replicate = DList.replicate
-    tailEx = DList.tail
-    {-# INLINE fromList #-}
-    {-# INLINE replicate #-}
-    {-# INLINE tailEx #-}
 
 instance SemiSequence (V.Vector a) where
     type Index (V.Vector a) = Int